app/pyproject.toml
[tool.poetry]
name = "b2broker"
version = "0.1.0"
description = "test task"
authors = ["max <megafukz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.2"
django-rest-framework = "^0.1.0"
pytest = "^8.0.0"
pytest-django = "^4.8.0"
mysqlclient = "^2.2.4"
dj-database-url = "^2.1.0"
python-dotenv = "^1.0.1"
django-filter = "^23.5"
drf-yasg = "^1.21.7"
factory-boy = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "^24.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
combine_as_imports = true
known_first_party = []
known_third_party = ['django', 'django_filters', 'ordered_model', 'rest_framework']
profile = "black"
skip = "migrations, templates"
skip_gitignore = true

[tool.black]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | manage.py           # also separately exclude a file named foo.py in the root of the project
)
'''
target-version = ['py311']

[pytest]
DJANGO_SETTINGS_MODULE = 'b2broker.settings'
python_files = ['test_*.py']
